move 'GuidGenerater' class to a separate file.
